requirements_multi_task_lfd.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argcomplete==3.0.8
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
boto==2.49.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.0.7
cosine-annealing-warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
crcmod==1.7
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.35
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
dm-tree==0.1.8
docker-pycreds==0.4.0
egl-probe==1.0.2
einops==0.6.1
executing==1.2.0
fasteners==0.18
fastjsonschema==2.17.1
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
fvcore @ git+https://github.com/facebookresearch/fvcore.git@fcc14bb0016408ddd818ddb0998f3df0e4e1381c
gcs-oauth2-boto-plugin==3.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.9
google-apitools==0.5.32
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-reauth==0.1.1
grad-cam==1.4.8
grpcio==1.54.2
gsutil==5.24
gym==0.21.0
gym-notices==0.0.8
h5py==3.8.0
httplib2==0.20.4
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
imageio==2.29.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kornia==0.6.12
learn2learn==0.2.0
llvmlite==0.40.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
model-index==0.1.11
monotonic==1.6
mujoco-py==2.1.2.14
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.4.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.0
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.17.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==2.6.9
pycocotools==2.0.6
pycparser==2.21
pygame==2.4.0
Pygments==2.15.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-ignite==0.4.12
pytorchvideo==0.1.5
pytz==2023.3
pyu2f==0.1.5
PyYAML==6.0
pyzmq==25.0.2
qpth==0.0.15
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
retry-decorator==1.1.1
rich==13.4.2
rsa==4.7.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.24.0
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1+cu117
torch-summary==1.4.5
torchcam==0.3.2
torchmetrics==0.11.4
torchvision==0.14.1+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
transforms3d==0.4.1
ttach==0.0.3
typing_extensions==4.6.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.3
wcwidth==0.2.6
Werkzeug==2.3.4
widgetsnbextension==4.0.7
yacs==0.1.8
yapf==0.40.0
yarl==1.9.2
zipp==3.15.0
